Refactor LLM modules with shared BaseLLM
