+++ src/deepthought/modules/llm_prod.py
@@ -1,26 +1,23 @@
-import json
 import logging
 import os
-from datetime import datetime, timezone
-from typing import List, Optional
+from typing import Optional
 
 import nats
-import torch
 from nats.aio.client import Client as NATS
-from nats.aio.msg import Msg
 from nats.js.client import JetStreamContext
 from peft import PeftModel
 from transformers import AutoModelForCausalLM, AutoTokenizer
 
 from ..config import get_settings
-from ..eda.events import EventSubjects, ResponseGeneratedPayload
+from ..eda.events import EventSubjects
 from ..eda.publisher import Publisher
 from ..eda.subscriber import Subscriber
+from .llm_base import BaseLLM
 
 logger = logging.getLogger(__name__)
 
 
-class ProductionLLM:
+class ProductionLLM(BaseLLM):
     """LLM module using a base model merged with LoRA adapter weights."""
 
     def __init__(
@@ -31,115 +28,22 @@ class ProductionLLM:
         adapter_dir: str = "./results/lora-adapter",
     ) -> None:
         model_name = model_name or get_settings().model_path
-        self._publisher = Publisher(nats_client, js_context)
-        self._subscriber = Subscriber(nats_client, js_context)
-        self._tokenizer = AutoTokenizer.from_pretrained(model_name)
+        publisher = Publisher(nats_client, js_context)
+        subscriber = Subscriber(nats_client, js_context)
+        tokenizer = AutoTokenizer.from_pretrained(model_name)
         base_model = AutoModelForCausalLM.from_pretrained(model_name)
         if os.path.isdir(adapter_dir):
             logger.info("Loading LoRA adapter from %s", adapter_dir)
             model = PeftModel.from_pretrained(base_model, adapter_dir)
-            self._model = model.merge_and_unload()
+            model = model.merge_and_unload()
         else:
             logger.warning(
                 "LoRA adapter directory %s not found. Using base model only.",
                 adapter_dir,
             )
-            self._model = base_model
+            model = base_model
+        super().__init__(publisher, subscriber, tokenizer, model)
         logger.info("ProductionLLM initialized with model %s", model_name)
-
-    def _build_prompt(self, facts: List[str]) -> str:
-        prompt = "\n".join(facts) + "\nResponse:" if facts else "Response:"
-        return prompt
-
-    async def _handle_memory_event(self, msg: Msg) -> None:
-        input_id = "unknown"
-        try:
-            data = json.loads(msg.data.decode())
-            if not isinstance(data, dict):
-                raise ValueError("MemoryRetrieved payload must be a dict")
-            input_id = data.get("input_id")
-            retrieved = data.get("retrieved_knowledge")
-            if not isinstance(input_id, str) or retrieved is None:
-                raise ValueError("Invalid memory payload fields")
-            if isinstance(retrieved, dict) and "retrieved_knowledge" in retrieved:
-                knowledge = retrieved.get("retrieved_knowledge", {})
-            elif isinstance(retrieved, dict):
-                knowledge = retrieved
-            else:
-                logger.error("retrieved_knowledge is not a dict for input_id %s", input_id)
-                if hasattr(msg, "nak") and callable(msg.nak):
-                    try:
-                        await msg.nak()
-                    except Exception:
-                        logger.error("Failed to NAK message", exc_info=True)
-                elif hasattr(msg, "ack") and callable(msg.ack):
-                    try:
-                        await msg.ack()
-                    except Exception:
-                        logger.error("Failed to ack message after error", exc_info=True)
-                return
-
-            facts = knowledge.get("facts")
-            if not isinstance(facts, list):
-                logger.error("retrieved_knowledge missing facts list for input_id %s", input_id)
-                if hasattr(msg, "nak") and callable(msg.nak):
-                    try:
-                        await msg.nak()
-                    except Exception:
-                        logger.error("Failed to NAK message", exc_info=True)
-                elif hasattr(msg, "ack") and callable(msg.ack):
-                    try:
-                        await msg.ack()
-                    except Exception:
-                        logger.error("Failed to ack message after error", exc_info=True)
-                return
-
-            logger.info("ProductionLLM received memory event ID %s", input_id)
-
-            prompt = self._build_prompt([str(f) for f in facts])
-            inputs = self._tokenizer(prompt, return_tensors="pt")
-            with torch.no_grad():
-                outputs = self._model.generate(**inputs, max_length=inputs["input_ids"].shape[1] + 20)
-            generated = self._tokenizer.decode(outputs[0], skip_special_tokens=True)
-            if generated.startswith(prompt):
-                response_text = generated[len(prompt) :].strip()  # noqa: E203
-            else:
-                response_text = generated.strip()
-
-            payload = ResponseGeneratedPayload(
-                final_response=response_text,
-                input_id=input_id,
-                timestamp=datetime.now(timezone.utc).isoformat(),
-                confidence=0.5,
-            )
-            await self._publisher.publish(EventSubjects.RESPONSE_GENERATED, payload, use_jetstream=True, timeout=10.0)
-            logger.info("ProductionLLM published RESPONSE_GENERATED for %s", input_id)
-            await msg.ack()
-        except (json.JSONDecodeError, ValueError) as e:  # pragma: no cover - validation errors
-            logger.error("Invalid MemoryRetrieved payload: %s", e, exc_info=True)
-            if hasattr(msg, "nak") and callable(msg.nak):
-                try:
-                    await msg.nak()
-                except Exception:
-                    logger.error("Failed to NAK message", exc_info=True)
-            elif hasattr(msg, "ack") and callable(msg.ack):
-                try:
-                    await msg.ack()
-                except Exception:
-                    logger.error("Failed to ack message after error", exc_info=True)
-
-        except Exception as e:  # pragma: no cover - runtime errors are logged
-            logger.error("Error in ProductionLLM handler: %s", e, exc_info=True)
-            if hasattr(msg, "nak") and callable(msg.nak):
-                try:
-                    await msg.nak()
-                except nats.errors.Error:
-                    logger.error("Failed to NAK message", exc_info=True)
-            elif hasattr(msg, "ack") and callable(msg.ack):
-                try:
-                    await msg.ack()
-                except nats.errors.Error:
-                    logger.error("Failed to ack message after error", exc_info=True)
 
     async def start_listening(self, durable_name: str = "llm_prod_listener") -> bool:
         if self._subscriber is None:
